Support a custom endpoint for mock environment
This is a small improvement to let a custom base URL be set when
the `Adyen::Client` is used in the `:mock` environment, which helps
when a different hostname or scheme is desired.

To help with compatibility, if both `mock_port` and `mock_service_url_base`
are specified, then the latter wins.